Changed Registration so that the name of the user in the database is the email address that they use.
Added fields to Registration for rewardScore, which keeps track of the total score progress towards getting
a reward and rewardLimit, which is the number of points needed to get a reward.

Added code in Task3_question.java to the block of code that fires when all 3 guesses are correct. The users
data for rewardScore is pulled from Firestore and one point is added, and then that new score is uploaded
back into Firestore.

Added code in Homepage.java that will pull the users current rewardScore from Firestore, and use that value
to fill in the progress bar at the top of the page. I figured out why Hannah had to add a button and couldn't
simply update the progress bar. It seems that Android does not read code in order like we are used to. Parts
of the code were being carried out before other parts, and the order matters for this progress bar. I added
code that delays the updating of the progress bar by 1 second to give all the other code time to load, and
time for Firebase to be queried for the current score of the user.

I added more bars to the progress bar so that we can display more than just 5 points. I have it set to 20
for now. This can be changed later if we decide to stick with this progress bar.